Fixed #13: Menu "Close" icon renders oddly in firefox.
Recent versions of Firefox (at least) seem to interpret the `transform-origin` property correctly (with regards to the element) when set in pixel values inside an SVG. When using keywords, they seem to be interpreted with regards to the SVG viewport instead.

Also, this commit simplifies the the rules for open/close icon, so that the "null transform" rules are removed and only the transformed state is specified. This also makes it necessary to hook the rule into the `.has-js` class set on the `html` element.